에디터 ì 추가, 이미지 업로드 코드 추가, Bible 조회 로직 ì수정

+++ src/main/java/theWordI/backend/domain/bible/dto/BibleBookResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class BibleBookResponse {
     private String versionId;
-    private Short bookId;
+    private int bookId;
     private String nameKo;
     private String testament;
     private Short chapterCount;
